refactor(python): refactoring upload file section
