+++ dialog_bot_sdk/bot.py
@@ -3,6 +3,7 @@ from .internal.bot import InternalBot
 from .entity_manager import EntityManager
 from .messaging import Messaging
 from .updates import Updates
+from .uploading import Uploading
 
 
 class DialogBot(object):
@@ -11,10 +12,10 @@ class DialogBot(object):
     """
     def __init__(self, channel, bot_token):
         self.internal = InternalBot(channel)
-        # TODO: analyze auth errors
         self.internal.authorize(bot_token)
         self.manager = EntityManager(self.internal)
         self.messaging = Messaging(self.manager, self.internal)
+        self.uploading = Uploading(self.internal)
         self.updates = Updates(self.manager, self.internal)
 
     @staticmethod
