+++ dialog_bot_sdk/utils/get_image_metadata.py
@@ -2,17 +2,17 @@ from PIL import Image
 import io
 
 
-def get_photo_w_h(file):
+def get_image_w_h(file):
     im = Image.open(file)
     return im.size
 
 
-def get_photo_thumb_bytes(file):
+def get_image_thumb_bytes(file):
     size = 90, 90
 
     im = Image.open(file)
     im = im.convert('RGB')
-    im.thumbnail(get_photo_w_h(file))
+    im.thumbnail(get_image_w_h(file))
     stream = io.BytesIO()
     im.save(stream, "JPEG")
